Add libbf

+++ src/common/rkmh.hpp
@@ -40,7 +40,7 @@ static char rev_arr[26] = {
         85, 86, 87, 88, 89, 90
 };
 
-typedef uint32_t hash_t;
+typedef uint64_t hash_t;
 
 std::vector<hash_t> hash_sequence(const char* seq,
                                   const uint64_t& len,
